requirements/base.txt
coupon-codes==0.1
Django==2.0.2
django-localflavor==2.0
django-phonenumber-field==2.0.0
psycopg2==2.7.3.2
